+++ packages/shells-bridge/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @minion-stack/shells-bridge
 
+## 0.1.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @minion-stack/shared@0.8.0
+
 ## 0.1.1
 
 ### Patch Changes